requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# Database Connections
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
elasticsearch==8.11.1

# Stream Processing
pyspark==3.5.0
kafka-python==2.0.2

# Data Generation
Faker==22.0.0
python-dateutil==2.8.2

# Visualization
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Monitoring & Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-elasticsearch==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-pymongo==0.42b0
opentelemetry-instrumentation-kafka==0.42b0
opentelemetry-exporter-otlp==1.21.0

# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
jupyter==1.0.0
jupyterlab==4.0.10
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Documentation
sphinx==7.1.2
sphinx-rtd-theme==2.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
tenacity==8.2.3

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Caching and Rate Limiting
redis==5.0.1
aioredis==2.0.1

# Telemetry and Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-sqlalchemy==0.42b0
prometheus-client==0.19.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 